requirements.txt
torch>=2.0.0
gymnasium[box2d]==0.29.1
wandb>=0.15.0
huggingface_hub>=0.19.0
moviepy>=1.0.3
pyyaml>=6.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0